Move integer arithmetic routines to separate file
